test-solution/requirements.txt
requests==2.25.1
pytest==6.2.5
PyHamcrest==2.0.2